requirements.txt
streamlit
openai-whisper
numpy
pickle5
chromadb
sentence-transformers
scikit-learn
pydub